+++ pkg/kt/command/exchange.go
@@ -1,9 +1,18 @@
 package command
 
 import (
-	"github.com/alibaba/kt-connect/pkg/kt/options"
+	"fmt"
+
 	"github.com/rs/zerolog"
+	"github.com/rs/zerolog/log"
+	
 	"github.com/urfave/cli"
+	"github.com/alibaba/kt-connect/pkg/kt/cluster"
+	"github.com/alibaba/kt-connect/pkg/kt/connect"
+	"github.com/alibaba/kt-connect/pkg/kt/options"
+	"github.com/alibaba/kt-connect/pkg/kt/util"
+
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 // newExchangeCommand return new exchange command
@@ -27,4 +36,43 @@ func newExchangeCommand(options *options.DaemonOptions) cli.Command {
 			return action.Exchange(c.Args().First(), options)
 		},
 	}
+}
+
+//Exchange exchange kubernetes workload
+func (action *Action) Exchange(swap string, options *options.DaemonOptions) error {
+	ch := SetUpCloseHandler(options)
+
+	checkConnectRunning(options.RuntimeOptions.PidFile)
+	expose := options.ExchangeOptions.Expose
+
+	if swap == "" || expose == "" {
+		return fmt.Errorf("-expose is required")
+	}
+
+	clientset, err := cluster.GetKubernetesClient(options.KubeConfig)
+	if err != nil {
+		return err
+	}
+
+	origin, err := clientset.AppsV1().Deployments(options.Namespace).Get(swap, metav1.GetOptions{})
+	if err != nil {
+		return err
+	}
+
+	replicas := origin.Spec.Replicas
+
+	// Prepare context inorder to remove after command exit
+	options.RuntimeOptions.Origin = swap
+	options.RuntimeOptions.Replicas = *replicas
+
+	factory := connect.Connect{}
+	_, err = factory.Exchange(options, origin, clientset, util.String2Map(options.Labels))
+	if err != nil {
+		return err
+	}
+
+	s := <-ch
+	log.Info().Msgf("Terminal Signal is %s", s)
+
+	return nil
 }
